services/scheduler/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pymysql 
databases 
alembic
aiomysql
async-exit-stack
async-generator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
celery[redis]
flower
sqlalchemy-celery-beat
mysql-connector
supabase
pyjwt
httpx
croniter==1.3.15
pytz
redis
chargebee
pyyaml